Add a CI job using the HttpBrowser of BrowserKit and HttpClient
