+++ tests/HttpKernelBrowserKitConfig.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Behat\Mink\Tests\Driver;
+
+use Behat\Mink\Driver\BrowserKitDriver;
+use Behat\Mink\Tests\Driver\Util\FixturesKernel;
+use Symfony\Component\HttpKernel\Client;
+
+class HttpKernelBrowserKitConfig extends AbstractBrowserKitConfig
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function createDriver()
+    {
+        $client = new Client(new FixturesKernel());
+
+        return new BrowserKitDriver($client);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getWebFixturesUrl()
+    {
+        return 'http://localhost';
+    }
+}
